Removed CatzRobotMap instances from autonomous paths. If catzrobotmap is instantiated multiple times within code, memory gets overwritten. Only need to instantiate once. Also changed reference to grabber within autonomous paths. AL 2/18/18

+++ Team-2637-PowerUp/src/autonomous/CatzPIDTurn.java
@@ -44,7 +44,7 @@ public class CatzPIDTurn
 	static boolean tuningMode = false;
 	static String debugData;
 	
-	public static void PIDturn(double degreesToTurn, int timeoutSeconds)
+	public static void PIDturn(double degreesToTurn, double timeoutSeconds)
 	{
 		boolean firstTime = true;
 		functionTimer = new Timer();
